Add Preferences option to automatically create a backup file (*.bak) when saving

+++ uk.ac.bolton.archimate.editor/src/uk/ac/bolton/archimate/editor/model/impl/EditorModelManager.java
@@ -45,6 +45,7 @@ import uk.ac.bolton.archimate.editor.Logger;
 import uk.ac.bolton.archimate.editor.diagram.util.AnimationUtil;
 import uk.ac.bolton.archimate.editor.model.IArchiveManager;
 import uk.ac.bolton.archimate.editor.model.IEditorModelManager;
+import uk.ac.bolton.archimate.editor.preferences.IPreferenceConstants;
 import uk.ac.bolton.archimate.editor.preferences.Preferences;
 import uk.ac.bolton.archimate.editor.ui.services.EditorManager;
 import uk.ac.bolton.archimate.editor.utils.FileUtils;
@@ -350,8 +351,8 @@ implements IEditorModelManager {
         
         File file = model.getFile();
         
-        // Save backup
-        if(file.exists()) {
+        // Save backup (if set in Preferences)
+        if(Preferences.STORE.getBoolean(IPreferenceConstants.BACKUP_ON_SAVE) && file.exists()) {
             FileUtils.copyFile(file, new File(model.getFile().getAbsolutePath() + ".bak"), false); //$NON-NLS-1$
         }
